OCR_KTP_API/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
openai
python-dotenv==1.0.0
Pillow==10.3.0
firebase-admin==6.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
boto3==1.34.116